CodeWhisper is a tool to do CODE discussions.
Doing code discussions is painful to refer to file paths line numbers
in EMail or messages.  These messages are difficult to read for
repeatedly switching between messages and source code.

With CodeWhisper, you review a WHOLE repository instead of a PR.  You
set a topic to discuss and add a list of comments on source files.
Send the discussion link to people you want him to join.  Participants
reply to comments and/or add new comments to do a discussion.